product_search.py added, TODO: implement the logger function
FIXME: searching, filtering and sorting works smoothly on Firefox and Chrome.
Safari throws a NoSuchFrameException.
Apparently, the product name is being entered correctly and the the search
button is also being clicked but nothing happends after that click.
Ideally, it should go to the searchresults page but it isn't going there.
(Only on Safari)